Organize into pkg and cmds directories
