+++ .circleci/config.yml
@@ -44,10 +44,10 @@ jobs:
       - run:
           name: vet
           command: |
-            go tool vet ./...
+            go tool vet cmds pkg
       - run:
           name: gofmt
-          command: test -z $(gofmt -s -l fmap integration pkg scripts uefi unicode utk utk2 uuid visitors)
+          command: test -z $(gofmt -s -l cmds pkg)
       - run:
           name: ineffassign
           command: ineffassign .
