+++ cmds/fmap/fake_test.flash
