+++ cmds/fmap/fmap.go
