+++ cmds/fmap/fmap_test.go
