+++ pkg/uefi/file_test.go
