+++ pkg/uefi/flash_test.go
