+++ pkg/uefi/flashdescriptormap.go
