+++ pkg/uefi/flashparams.go
