+++ pkg/uefi/flashregionsection.go
