+++ pkg/uefi/gberegion.go
