+++ pkg/uefi/region.go
