+++ pkg/uefi/region_test.go
