+++ pkg/unicode/ucs2.go
