+++ pkg/visitors/extract_test.go
@@ -10,7 +10,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/linuxboot/fiano/uefi"
+	"github.com/linuxboot/fiano/pkg/uefi"
 )
 
 var (
@@ -21,7 +21,7 @@ var (
 
 func init() {
 	var err error
-	sampleFV, err = ioutil.ReadFile("../integration/roms/ovmfSECFV.fv")
+	sampleFV, err = ioutil.ReadFile("../../integration/roms/ovmfSECFV.fv")
 	if err != nil {
 		log.Fatal(err)
 	}
