+++ pkg/visitors/parsedir.go
@@ -9,7 +9,7 @@ import (
 	"io/ioutil"
 	"os"
 
-	"github.com/linuxboot/fiano/uefi"
+	"github.com/linuxboot/fiano/pkg/uefi"
 )
 
 // ParseDir creates the firmware tree and reads the binaries from the provided directory
