+++ pkg/visitors/replacepe32.go
@@ -4,7 +4,7 @@ import (
 	"io/ioutil"
 	"regexp"
 
-	"github.com/linuxboot/fiano/uefi"
+	"github.com/linuxboot/fiano/pkg/uefi"
 )
 
 // ReplacePE32 replaces PE32 sections with NewPE32 for all files matching Predicate.
